Create pseudorandom_number_generator
Made a pseudorandom_number_generator (PRNG) to learn how they work and see if I could make one. This one is based on a Linear Congruential Generator formula but is modified after I noticed the numbers would repeat. 

It takes an input and the amount of output random numbers you want to get and outputs the random numbers. If graphed on a scatterplot you will see they are truly random.